Dramatically improved string rendering, fixed multiblocks, and more

+++ Classes/Buffer.cpp
@@ -23,11 +23,14 @@ void Buffer::Create(const std::vector<int> &config, const Data &data) {
     glBindBuffer(GL_ARRAY_BUFFER, VBO);
 
     if (data.size() > 0) {
-        glBufferData(GL_ARRAY_BUFFER, data.size() * sizeof(float), data.data(), GL_STATIC_DRAW);
+        glBufferData(
+            GL_ARRAY_BUFFER, static_cast<long>(data.size() * sizeof(float)),
+            data.data(), GL_STATIC_DRAW
+        );
         Vertices = static_cast<int>(data.size()) / VertexSize;
     }
 
-    int index = 0;
+    unsigned int index = 0;
     int partSum = 0;
 
     for (int const &element : config) {
@@ -47,10 +50,16 @@ void Buffer::Upload(const Data &data, int start, bool sub) {
 
     if (start == 0 && !sub) {
         Vertices = static_cast<int>(data.size()) / VertexSize;
-        glBufferData(GL_ARRAY_BUFFER, data.size() * sizeof(float), data.data(), GL_STATIC_DRAW);
+        glBufferData(
+            GL_ARRAY_BUFFER, static_cast<long>(data.size() * sizeof(float)),
+            data.data(), GL_STATIC_DRAW
+        );
     }
     else {
-        glBufferSubData(GL_ARRAY_BUFFER, start, data.size() * sizeof(float), data.data());
+        glBufferSubData(
+            GL_ARRAY_BUFFER, start,
+            static_cast<long>(data.size() * sizeof(float)), data.data()
+        );
     }
 
     glBindBuffer(GL_ARRAY_BUFFER, 0);
@@ -74,7 +83,7 @@ void Buffer::Draw(int start, int length) {
     BufferShader->Unbind();
 }
 
-void UniformBuffer::Create(std::string name, int bufferID, int size, std::vector<Shader*> shaders) {
+void UniformBuffer::Create(std::string name, unsigned int bufferID, int size, std::vector<Shader*> shaders) {
     BufferID = bufferID;
 
     glGenBuffers(1, &UBO);
@@ -90,19 +99,22 @@ void UniformBuffer::Create(std::string name, int bufferID, int size, std::vector
     glBindBufferRange(GL_UNIFORM_BUFFER, 0, UBO, 0, size);
 }
 
-void UniformBuffer::Add(std::string name, int bufferID, Shader* shader) {
+void UniformBuffer::Add(std::string name, unsigned int bufferID, Shader* shader) {
     glUniformBlockBinding(shader->Program, glGetUniformBlockIndex(shader->Program, name.c_str()), bufferID);
 }
 
 template <typename T>
-void UniformBuffer::Upload(int index, T t) {
+void UniformBuffer::Upload(unsigned long index, T t) {
     glBindBuffer(GL_UNIFORM_BUFFER, UBO);
     glBufferSubData(GL_UNIFORM_BUFFER, sizeof(T) * index, sizeof(T), t);
     glBindBuffer(GL_UNIFORM_BUFFER, 0);
 }
 
-void UniformBuffer::Upload(int index, glm::mat4 matrix) {
+void UniformBuffer::Upload(unsigned long index, glm::mat4 matrix) {
     glBindBuffer(GL_UNIFORM_BUFFER, UBO);
-    glBufferSubData(GL_UNIFORM_BUFFER, sizeof(glm::mat4) * static_cast<unsigned long>(index), sizeof(glm::mat4), glm::value_ptr(matrix));
+    glBufferSubData(
+        GL_UNIFORM_BUFFER, static_cast<long>(sizeof(glm::mat4) * index),
+        sizeof(glm::mat4), glm::value_ptr(matrix)
+    );
     glBindBuffer(GL_UNIFORM_BUFFER, 0);
 }
